Fixed compilation for MacOS X and gcc 4.8! (Shankar from Colin's computer)

+++ jdftx/core/DataCollection.h
@@ -89,7 +89,7 @@ template<typename T> void loadFromFile(TptrCollection& x, const char* filename)
 	if(fLen != expectedLen)
 	{	die("\nLength of '%s' was %ld instead of the expected %ld bytes.\n"
 				"Hint: Are you really reading the correct file?\n\n",
-				filename, fLen, expectedLen);
+				filename, (unsigned long)fLen, (unsigned long)expectedLen);
 	}
 	
 	FILE* fp = fopen(filename, "rb");
