+++ jdftx/core/DataIO.h
@@ -61,7 +61,7 @@ template<typename T> void loadRawBinary(Tptr& X, const char* filename)
 	if(fLen != expectedLen)
 	{	die("\nLength of '%s' was %ld instead of the expected %ld bytes.\n"
 				"Hint: Are you really reading the correct file?\n\n",
-				filename, fLen, expectedLen);
+				filename, (unsigned long)fLen, (unsigned long)expectedLen);
 	}
 	
 	loadRawBinary(X, fp);
